requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Reinforcement Learning
gymnasium>=0.29.0
pettingzoo>=1.24.0
stable-baselines3>=2.0.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Processing & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0

# Image/Animation Generation
imageio>=2.28.0
pillow>=9.5.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
